% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acc}
\alias{acc}
\title{A accuracy computation function}
\usage{
acc(N, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{a}{matrix a}

\item{b}{matrix b}
}
\value{
the accuracy ratio
}
\description{
A accuracy computation function using Rcpp
}
\examples{
\dontrun{
acc(3,as.matrix(c(1,2,3)),as.matrix(c(2,3,4)))
}
}
