% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{cnn}
\alias{cnn}
\title{CNN network used for classification}
\usage{
cnn(datasets, num_classes, batch_size, epochs)
}
\arguments{
\item{datasets}{the dataset used for training}

\item{num_classes}{the total kinds of labels}

\item{batch_size}{the size of samples will be used to update CNN at one time}

\item{epochs}{number of CNN training times}
}
\value{
the accurate rate and prediction results of the trained model on test set
}
\description{
CNN network for multi-class graph classification
}
\examples{
\dontrun{
num_classes <- 10
batch_size <- 128
epochs <- 10
datasets = dataset_mnist()
cnn(datasets,num_classes,batch_size,epochs)
}
}
