\name{StatComp22080-package}
\alias{StatComp22080-package}
\alias{StatComp21080}
\docType{package}
\title{
  CNN network used for classification
}
\description{
  This package provides CNN network for multi-class graph classification.
}
\details{
  R functions of CNN network is builded by Keras, it contains two convolution layers followed by max-pooling layer and two blocks of drop-out layer combined with dense layer, the loss function is crossentropy and optimizer is rmsprop.
}
\author{
Cai Yide

Maintainer: Cai Yide <sa21229001@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
